backend/requirements.txt
# API Framework
fastapi>=0.110.0
uvicorn[standard]>=0.27.1

# HTTP client
httpx>=0.27.0

# Testing
pytest>=8.0.2

# Linting
flake8>=7.0.0

# Database driver (para Semana 2)
psycopg2-binary>=2.9.9

# Pydantic (ya viene con FastAPI, pero explícito)
pydantic>=2.6.0

# Seguridad (Día 10)
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9

# Migraciones (Día 11)
yoyo-migrations>=8.2.0

# BD — Driver PostgreSQL (Día 14)
# psycopg3 reemplaza a psycopg2 por compatibilidad con Python 3.13 en Windows
psycopg[binary]>=3.1.0